Fix bug in scaling of hands (added an additional factor of "animal_size")
Corrected in rendering of the animals and in the parameter values of the animal parameter .mat files.  'hand length' parameter now, analogously with 'foot length' parameter, specifies the length of the hand relative to the size of the animal, where as before it was relative to the size of the animal with an additional factor of animal size and foreleg radius.