Added generalization stimuli.
These were generated by making an "X" in the original two-dimensional stimulus subspace, plus a third, orthogonal direction.  The .mat file contains the parameters used to generate these, including:
 * the origin and two original direction (from disc 8)
 * a third, orthogonal direction (dir1 from disc 9)
 * the unit-length direction vectors, and the distances along those vectors (relative to the distance of the in-category items from the prototype; 0.0625 in this case)
 * the xy coordinates for the two in-space generalization directions (use xy_to_animal_image with dir1 and dir2)
 * the xy coordinates for the new direction (use xy_to_animal_image with dir1 and dir3)
 * the xy coordinates of the prototype, which is the same as the subspace origin.

The png file shows the resulting images from these parameters, with the prototype in the middle column and the two original direction and the new direction in the three rows.