Factor out code for generating stimuli from stim_script into make_stimuli; more easier stimulus sets
The original stimulus set had a distance from the prototype for the "out of category" stimuli that was twice that of the inner category ones, but this meant that the inner and outer rings were very close and the task was very hard.  I generated sets with a 3:1 and 4:1 ratio.  None of these stimuli look to extreme to my eyes.