Fix foot size issue.
For some of the more extreme animals the foot size could have been used as a single feature cue to category ID, so there are now versions of the stimulus sets without the variation in foot size. Foot eccentricity still varies, though, but it's not as salient as far as I can tell.